Refactor config handling and reorganize code structure
- Rename cfg to tiledb in context object for clarity
- Create new utils/cfg.py module for centralized configuration management
- Move metadata utilities from cli/metadata/utils.py to utils/metadata.py
- Enhance Dask deployment checks with specific deployment types
- Update import statements across codebase to reflect new structure
