+++ src/gwasstudio/utils/metadata.py
@@ -9,7 +9,6 @@ from ruamel.yaml import YAML
 from gwasstudio import logger
 from gwasstudio.mongo.models import EnhancedDataProfile, DataProfile
 from gwasstudio.utils import lower_and_replace, compute_sha256
-from gwasstudio.utils.vault import get_config_from_vault
 
 metadata_dtypes = {"project": "category", "study": "category", "file_path": "string[pyarrow]", "category": "category"}
 
@@ -199,12 +198,3 @@ def ingest_metadata(df: pd.DataFrame, mongo_uri: str = None) -> None:
     for document in documents:
         obj = EnhancedDataProfile(uri=mongo_uri, **document)
         obj.save()
-
-
-def get_mongo_uri(ctx: object) -> str:
-    """Retrieve MongoDB URI from Vault or command line options."""
-
-    vault_options = ctx.obj.get("vault")
-    mongo_config = get_config_from_vault("mongo", vault_options)
-
-    return mongo_config.get("uri") or ctx.obj.get("mongo").get("uri")
